# internal/api/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D2R Offline Farming Bot Local Core API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "0"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v1/history/delete-all/preview": {
      "post": {
        "operationId": "previewHistoryDeleteAll",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/HistoryDeletePreviewRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metadata-bound delete-all preview",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HistoryDeletePreviewDTO"}
              }
            }
          }
        }
      }
    },
    "/api/v1/history/delete-all/confirm": {
      "post": {
        "operationId": "confirmHistoryDeleteAll",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/HistoryDeleteConfirmRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirmed delete-all result",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/HistoryDeleteResultDTO"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "HistoryMaintenanceDiagnosticDTO": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "file_id": {"type": "string"},
          "code": {"type": "string"}
        }
      },
      "HistoryDeletePreviewRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_generation"],
        "properties": {
          "expected_generation": {"type": "integer"}
        }
      },
      "HistoryDeletePreviewDTO": {
        "type": "object",
        "required": ["confirmation_token", "index_generation", "candidate_files", "candidate_bytes", "protected_files", "categories"],
        "properties": {
          "confirmation_token": {"type": "string"},
          "index_generation": {"type": "integer"},
          "candidate_files": {"type": "integer"},
          "candidate_bytes": {"type": "integer"},
          "protected_files": {"type": "integer"},
          "categories": {"type": "object", "additionalProperties": {"type": "integer"}}
        }
      },
      "HistoryDeleteConfirmRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expected_generation", "confirmation_token", "index_generation", "candidate_files", "candidate_bytes"],
        "properties": {
          "expected_generation": {"type": "integer"},
          "confirmation_token": {"type": "string"},
          "index_generation": {"type": "integer"},
          "candidate_files": {"type": "integer"},
          "candidate_bytes": {"type": "integer"}
        }
      },
      "HistoryDeleteResultDTO": {
        "type": "object",
        "required": ["deleted_files", "deleted_bytes", "protected_files", "diagnostics"],
        "properties": {
          "deleted_files": {"type": "integer"},
          "deleted_bytes": {"type": "integer"},
          "protected_files": {"type": "integer"},
          "diagnostics": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/HistoryMaintenanceDiagnosticDTO"}
          }
        }
      },
...
    }
  }
}